Mms: add MobilePaperShowActivity
This activity is from CAF, which allows incoming MMS messages which
appear as slide shows to be broken up into their separate images, while
still exposing the slideshow functinoality via an overflow menu button.

This helps alleviate the symptom of when an iPhone will send more than
one image in one MMS message and Mms app interpreting it as a slideshow
instead of separate images.

This commit is essentially a squash of:
https://www.codeaurora.org/cgit/quic/la/platform/packages/apps/Mms/log/src/com/android/mms/ui/MobilePaperShowActivity.java?h=LNX.LA.3.7.1&showmsg=1

Change-Id: I131a7b6203a071fc4adfc0efc80811a1e51aa995

Conflicts:

	res/values/cm_strings.xml